+++ lib/core/error/exceptions.dart
@@ -1,0 +1,16 @@
+// lib/core/error/exceptions.dart
+
+class ServerException implements Exception {
+  final String message;
+  const ServerException(this.message);
+}
+
+class DatabaseException implements Exception {
+  final String message;
+  const DatabaseException(this.message);
+}
+
+class CacheException implements Exception {
+  final String message;
+  const CacheException(this.message);
+}