+++ lib/features/expense/domain/usecases/get_expense.dart
@@ -1,0 +1,12 @@
+import '../entities/entities.dart';
+import '../repositories/expense_repository.dart';
+
+class GetExpenseUseCase {
+  final ExpenseRepository repository;
+
+  const GetExpenseUseCase(this.repository);
+
+  Future<void> call() async {
+    // TODO: implement
+  }
+}