+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0+1 (10 Jun 2026)
 
 ### feat
+- add core expense category
 - add package dependencies for pubspec.yaml
 - add scripts for generated code folder featured
 - initial commit